Add vim-textobj-entire and vim-plug to ease actions on entire buffer
I wanted to make it easier to replace the entire text of a buffer
with something on my clipboard and ggVGp didn't work very well. With
this plugin (which meant I had to install the vim-plug plugin system)
I can now write Vaep instead and it replaces the entire text of the
buffer with whatever is in the clipboard exactly as Ctrl-A Ctrl-V
would do in a conventional text editor (which means said editor wins
out over vim in this aspect).