Add vim-modusline plugin, configure its color settings
I installed the vim-modusline plugin, as I thought changing the color
of the statusline depending on mode was nifty, but didn't want to use
a statusline plugin (like lightline), given I use a custom statusline.
I didn't like all of the default settings however, so I changed some,
using an autocommand script provided by the author of the plugin.