Delete vimrc-old file
I noticed this file became a year old yesterday, and I've changed a
lot of things during that time, so I decided to remove it entirely. If
anyone wants to see my current vimrc, it remains right there in the
repo, and also the old vimrc file can still be accessed by the git history.