:wrench: Restore BufWinEnter method
Turns out that using FileType help has the consequence that when
accessing help with SkyBison later in a session (as opposed to the
first time) the help buffer fails to cover the entire viewpane. Seems
that this event activates only the first time a file of the specific
filetype is accessed.