:wrench: Replace SyntaxAttr.vim with inspecthi.vim
SyntaxAttr couldn't be called from the command line (the area with the
semicolon) without using :call and had to be bound to a key. inspecthi
can be called like that just fine and thus doesn't need to be bound to a
key. Another difference is that SyntaxAttr also showed the properties of
the element under cursor (color and emphasis), and inspecthi only shows
the groups where the current element inherits from. Still I preferred
using inspecthi, even from the time it was part of pgmnt, so I'm glad it
is now packed by itself.